+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- *chore*: add sentinel example
+
 ## [1.0.0-rc.6] - 2026-04-09
 
 - *bump*: Bump to v1.0.0-rc.7 (#58)